moved AddLayer scripts and LayerHierarchyInspector scripts
