+++ Assets/_Application/Layers/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d9deaefb4f778405e9857e777644caa8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
