+++ Assets/_Application/Layers/UI/AddLayer/AddLayerPanel.cs
@@ -1,11 +1,9 @@
-using System;
 using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Events;
 using UnityEngine.UI;
 
-namespace Netherlands3D.Twin
+namespace Netherlands3D.Twin.Layers.UI.AddLayer
 {
     public class AddLayerPanel : MonoBehaviour
     {
