+++ Assets/_Application/Layers/UI/AddLayer/LayerToolBarButtonAddFolder.cs
@@ -1,11 +1,6 @@
-using System.Collections;
-using System.Collections.Generic;
-using Netherlands3D.Twin.Projects;
-using Netherlands3D.Twin.UI.LayerInspector;
-using UnityEngine;
 using UnityEngine.EventSystems;
 
-namespace Netherlands3D.Twin
+namespace Netherlands3D.Twin.Layers.UI.AddLayer
 {
     public class LayerToolBarButtonAddFolder : LayerToolBarButtonBase
     {
