+++ Assets/_Application/Layers/UI/AddLayer/LayerToolBarButtonDeleteLayer.cs
@@ -1,10 +1,6 @@
-using System.Collections;
-using System.Collections.Generic;
-using Netherlands3D.Twin.UI.LayerInspector;
-using UnityEngine;
 using UnityEngine.EventSystems;
 
-namespace Netherlands3D.Twin
+namespace Netherlands3D.Twin.Layers.UI.AddLayer
 {
     public class LayerToolBarButtonDeleteLayer : LayerToolBarButtonBase
     {
