+++ Assets/_Application/Layers/UI/AddLayer/LayerToolBarToggleAddLayer.cs
@@ -1,9 +1,7 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.EventSystems;
 
-namespace Netherlands3D.Twin
+namespace Netherlands3D.Twin.Layers.UI.AddLayer
 {
     public class LayerToolBarToggleAddLayer : LayerToolBarToggleBase
     {
