+++ Assets/_Application/Layers/UI/HierarchyInspector.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cc946c435a2c549bb99a0efb964a5935
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
