+++ Assets/_Functionalities/OGC3DTiles/Scripts/Tile3DLayerToggle.cs
@@ -1,5 +1,5 @@
 using System.Linq;
-using Netherlands3D.Twin.UI.LayerInspector;
+using Netherlands3D.Twin.Layers.UI.AddLayer;
 using UnityEngine;
 
 namespace Netherlands3D.Functionalities.OGC3DTiles
