+++ Assets/_Functionalities/OGC3DTiles/Scripts/Tiles3DContentOverlayInspector.cs
@@ -1,6 +1,7 @@
 using Netherlands3D.Twin;
 using Netherlands3D.Twin.Layers;
 using Netherlands3D.Twin.Layers.LayerTypes.Credentials.Properties;
+using Netherlands3D.Twin.Layers.UI.HierarchyInspector;
 using Netherlands3D.Twin.UI.LayerInspector;
 using UnityEngine;
 using UnityEngine.Networking;
