+++ Assets/_Functionalities/UrbanReLeaf/Scripts/SensorLayer/SensorLayerToggle.cs
@@ -1,4 +1,5 @@
-using Netherlands3D.Twin.UI.LayerInspector;
+using Netherlands3D.Twin.Layers.UI.AddLayer;
+using Netherlands3D.Twin.Layers.UI.HierarchyInspector;
 
 namespace Netherlands3D.Functionalities.UrbanReLeaf
 {
